Move Lambda Lab tests into smoke tests and change local tag file location
